Rename term.{h,c} to textdisp.{h,c} #52
